requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
alembic==1.14.0
    # via test (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   fast-depends
    #   starlette
    #   watchfiles
asyncmy==0.2.9
    # via test (pyproject.toml)
certifi==2024.8.30
    # via http3
chardet==3.0.4
    # via http3
click==8.1.7
    # via uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via test (pyproject.toml)
fast-depends==2.4.12
    # via test (pyproject.toml)
fastapi==0.115.5
    # via test (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.8.1
    # via
    #   http3
    #   uvicorn
h2==3.2.0
    # via http3
hpack==3.0.0
    # via h2
http3==0.6.7
    # via test (pyproject.toml)
httptools==0.6.4
    # via uvicorn
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   anyio
    #   email-validator
    #   http3
iniconfig==2.0.0
    # via pytest
mako==1.3.6
    # via alembic
markupsafe==3.0.2
    # via mako
orjson==3.10.11
    # via test (pyproject.toml)
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.10
    # via test (pyproject.toml)
pydantic==2.9.2
    # via
    #   fast-depends
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.1
    # via test (pyproject.toml)
pytest==8.3.3
    # via
    #   pytest-asyncio
    #   pytest-mock
    #   test (pyproject.toml)
pytest-asyncio==0.24.0
    # via test (pyproject.toml)
pytest-mock==3.14.0
    # via test (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-ulid==3.0.0
    # via test (pyproject.toml)
pyyaml==6.0.2
    # via uvicorn
rfc3986==1.5.0
    # via http3
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.36
    # via
    #   alembic
    #   sqlalchemy-utils
    #   sqlmodel
    #   test (pyproject.toml)
sqlalchemy-utils==0.41.2
    # via test (pyproject.toml)
sqlmodel==0.0.22
    # via test (pyproject.toml)
starlette==0.41.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn[standard]==0.32.0
    # via test (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==14.1
    # via uvicorn
